% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmbox.r
\name{csmbox}
\alias{csmbox}
\title{csmbox
Display expression of a selected gene}
\usage{
library(hscsm)

data(csm)

csmbox(csm,gene="RUNX2")
}
\description{
csmbox
Display expression of a selected gene
}
\examples{
library(hscsm)
data(csm)
csmbox(csm,gene="RUNX2")
}
